Fix res sign

+++ packages/db/clibs/lib/selva/selva.c
@@ -882,7 +882,7 @@ static napi_value selva_find(napi_env env, napi_callback_info info)
     };
 
     err = find(db, node, &cb_wrap);
-    return res2napi(env, err ?: ((struct selva_find_cb *)cb_wrap.node_arg)->nr_results);
+    return res2napi(env, err ?: (int)((struct selva_find_cb *)cb_wrap.node_arg)->nr_results);
 }
 
 // selva_save(db, filename): number
